Uploaded final assessment and the powerpoint used during presentation.
I followed up with this presentation by teaching the class about Bollinger Bands and how they are useful to determine if something is in fact affecting your stock. I used Bollinger bands to do just that and I was willing to wager it was the finding from my fundamental analysis.